+++ src/test/java/com/example/mohago_nocar/plan/presentation/TravelPlanControllerV2Test.java
@@ -1,5 +1,6 @@
 package com.example.mohago_nocar.plan.presentation;
 
+import com.example.mohago_nocar.plan.presentation.v2.TravelPlanControllerV2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
